Swap bundler and add some initial unit tests.
I changed the bundler from webpack to vite. I'll need to investigate
what is required to support legacy browsers, but for now, prefer the new
hotness. Part of this decision was based on it's ease of integration
with vitest, which is very similar to jest. It also avoids a strange
warning from openpgp during tests.