Add more tests and tune vite import globs.
Some of the barrel index files were using globs that included
everything in their respective directory. This included the test files
and caused the bundler to go bonkers. I had to modify the glob to
exclude them. I also decided to make the access tokens for providers
required to encourage safer usage. The user should still make the repos
private.