More testing.
Found a weird issue with vitest where mocked stores were not showing the
correct values if the base url was set in the vite config. I unset the
option in the config file for now since it's not being used. I also added a
specific test for when the base url is set, then changed the other tests
to not assume a base url is set. This seems cleaner, anyway, and I think
the issue can be worked around.